+++ src/test/java/lotto/domain/vo/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
